feat: Refactor folder and user management services for improved search functionality and code consistency

+++ DocumentManagementUI/src/app/admin/usercard/usercard.component.ts
@@ -1,6 +1,8 @@
 import { DatePipe } from '@angular/common';
 import { Component, inject, OnInit } from '@angular/core';
 import { UsercardserviceService } from './usercardservice.service';
+import { tap } from 'rxjs';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-usercard',
@@ -13,11 +15,10 @@ export class UsercardComponent implements OnInit {
   private usercardservice = inject(UsercardserviceService);
 
   deleteUser(userId: number) {
-    this.usercardservice.deleteUser(+userId).subscribe(
+    this.usercardservice.deleteUser(+userId).pipe(tap(this.users = this.users.filter((user: any) => user.id !== userId))).subscribe(
       {
         next: (response) => {
           console.log('User deleted successfully:', response);
-          this.users = this.users.filter((user: any) => user.id !== userId);
         },
         error: (error) => {
           console.error('Error deleting user:', error);
@@ -27,14 +28,13 @@ export class UsercardComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.usercardservice.getAllUsers().subscribe(
-      (response) => {
+    this.usercardservice.getAllUsers().subscribe({
+      next: (response) => {
         this.users = response.item1;
       },
-      (error) => {
+      error: (error) => {
         console.error('Error fetching users:', error);
       }
-    );
+    });
   }
-
 }
